requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
numpy>=1.26.0
python-dotenv>=1.0.0